+++ config.lua
@@ -1,0 +1,2 @@
+---------------
+EASY_STARTUP = settings.global["UnRealistic-easy-startup"].value