+++ settings.lua
@@ -1,0 +1,8 @@
+data:extend({
+	{
+		type = "bool-setting",
+		name = "UnRealistic-easy-startup",
+		setting_type = "runtime-global",
+		default_value = false
+	},
+})